requirements.txt
Pillow==7.1.2
PyPDF2~=1.26.0